Added IPFIX support and further improvements

+++ src/flowwriter.rs
@@ -22,7 +22,6 @@ pub struct FlowWriter {
     base_dir    : String,
     threshold   : u64,              // Number of flows collected that will lead to a rotation
     schema      : Schema,           // Schema of the parquet file
-    writer      : ArrowWriter<std::fs::File>,
     flows       : Vec<FlowStats>,   // Will contain flowstats
     last_rot    : Instant,          // Set when rotated, helps to determine if flush needed at next tick 
     insert_ch   : bool,             // Will be set to true if a db_table is given
@@ -48,27 +47,12 @@ impl FlowWriter {
     {
         let fields = FlowStats::create_fields();
         let schema = Schema::new(fields.clone());
-        let filename = format!("{}/flows.current.parquet", base_dir);
-        let props = WriterProperties::builder()
-        .set_writer_version(WriterVersion::PARQUET_2_0)
-        .set_encoding(Encoding::PLAIN)
-        .set_compression(Compression::SNAPPY)
-        .build();
-    // .set_column_encoding(ColumnPath::from("col1"), Encoding::DELTA_BINARY_PACKED)
-    
-        // eprintln!("Trying to open file {}", filename);
-        let file =  File::create(filename.clone()).unwrap();
-
-        let writer = 
-            ArrowWriter::try_new(file, Arc::new(schema.clone()), Some(props))
-            .expect("Error opening parquet file");
    
         let mut fw = FlowWriter {
             rx          : rx,
             base_dir    : base_dir.to_string(),
             threshold   : threshold,
             schema      : schema.clone(),
-            writer      : writer,
             flows       : Vec::new(),
             last_rot    : Instant::now(),
             insert_ch   : ch_db_table.is_some(),
@@ -117,30 +101,9 @@ impl FlowWriter {
             let delta: f64 = self.last_rot.elapsed().as_millis() as f64;
             debug!("Buffer threshold reached. Rotating. {:.1} kflows/s", self.threshold as f64/delta);
             // self.write_batch();
-            self.rotate(true);
+            self.rotate();
         }
 
-    }
-
-    fn write_batch(&mut self) {
-        let batch = self.record_batch();
-        debug!("Writing batch ({} flows)", batch.num_rows());
-        self.writer.write(&batch).unwrap();
-        self.flows = Vec::new();
-    }
-
-
-    fn close_current(&mut self) {
-        // self.flush();
-        // write any remaining flows
-        self.write_batch();
-        let _ = self.writer.flush();
-        match self.writer.finish() {
-            Err(error) => {
-                error!("Error closing writer : {:?}", error);
-            }
-            _ => (),
-        }
     }
 
     fn rotate_tick(&mut self, open_new: bool) {
@@ -149,44 +112,47 @@ impl FlowWriter {
 
         if self.last_rot.elapsed() > Duration::from_secs(2) || !open_new {
             if self.flows.len() > 0 {
-                self.rotate(open_new);
-            } else {
-                // remove the 'current' file
-                let filename = format!("{}/flows.current.parquet", self.base_dir.clone());
-                if !open_new {
-                    let _ = std::fs::remove_file(filename);
-                }
+                self.rotate();
             }
         } else {
             debug!("Last rotation less than 2 seconds ago, not bothered");
         }
     }
 
-    fn rotate(&mut self, open_new:bool) {
+    fn rotate(&mut self) {
         // reset timer
         self.last_rot = Instant::now();
         // Close current, rename, open new
-        self.close_current();
+        // self.close_current();
         let loc_now: DateTime<Local> = Local::now();
-        let filename = format!("{}/flows.current.parquet", self.base_dir.clone());
+        // let filename = format!("{}/flows.current", self.base_dir.clone());
         // Rename to naming scheme
         let to_file = format!("{}/flows-{}.parquet", self.base_dir.clone(), loc_now.format("%Y-%m-%d-%H:%M:%S%.6f"));
-        debug!("rename {} -> {}", filename.clone(), to_file.clone());
-        std::fs::rename(filename.clone(), to_file.clone()).unwrap();
-        // Create new one if requested
-        if open_new {
-            let props = WriterProperties::builder()
+        debug!("Writing to {}", to_file.clone());
+        // std::fs::rename(filename.clone(), to_file.clone()).unwrap();
+        let props = WriterProperties::builder()
             .set_writer_version(WriterVersion::PARQUET_2_0)
             .set_encoding(Encoding::PLAIN)
             .set_compression(Compression::SNAPPY)
             .build();
-        // .set_column_encoding(ColumnPath::from("col1"), Encoding::DELTA_BINARY_PACKED)
-        
-            // eprintln!("Trying to open file {}", filename);
-            let file = File::create(filename).unwrap();
-            self.writer = ArrowWriter::try_new(file, Arc::new(self.schema.clone()), Some(props)).unwrap();
-        }
+        // eprintln!("Trying to open file {}", filename);
+        let file = File::create(to_file.clone()).unwrap();
+        let mut writer = ArrowWriter::try_new(file, Arc::new(self.schema.clone()), Some(props)).unwrap();
     
+        let batch = self.record_batch();
+        debug!("Writing batch ({} flows)", batch.num_rows());
+        writer.write(&batch).unwrap();
+        let _ = writer.flush();
+        match writer.finish() {
+            Err(error) => {
+                error!("Error closing writer : {:?}", error);
+            }
+            _ => (),
+        }
+
+        self.flows = Vec::new();
+
+        
         // If we have a database then insert it into database and remove the rotated file
         if self.insert_ch {
             // Setting stdin to this file does not always work for some reason
@@ -195,12 +161,11 @@ impl FlowWriter {
             let query = format!("INSERT INTO {} FROM INFILE '{}' FORMAT Parquet", 
                 self.ch_db_table.as_ref().clone().unwrap(),
                 to_file.clone());
-            self.ch_query(query);
-
-            // If we insert into CH, then we delete the file afterwards
-            std::fs::remove_file(to_file.clone()).unwrap();
+            if self.ch_query(query) {
+                // If we insert into CH succesfully, then we delete the file afterwards
+                std::fs::remove_file(to_file.clone()).unwrap();
+            }
         }
-
     }
 
     fn record_batch(&mut self) -> RecordBatch {
@@ -212,8 +177,8 @@ impl FlowWriter {
         let dp = UInt16Array::from(self.flows.iter().map(|p| p.dp).collect::<Vec<Option<u16>>>());
         let pr = GenericStringArray::<i32>::from(self.flows.iter().map(|p| p.pr.clone()).collect::<Vec<Option<String>>>());
         let flg = GenericStringArray::<i32>::from(self.flows.iter().map(|p| p.flg.clone()).collect::<Vec<Option<String>>>());
-        // let icmp_type = UInt8Array::from(self.flows.iter().map(|p| p.icmp_type).collect::<Vec<Option<u8>>>());
-        // let icmp_code = UInt8Array::from(self.flows.iter().map(|p| p.icmp_code).collect::<Vec<Option<u8>>>());
+        let icmp_type = UInt8Array::from(self.flows.iter().map(|p| p.icmp_type).collect::<Vec<Option<u8>>>());
+        let icmp_code = UInt8Array::from(self.flows.iter().map(|p| p.icmp_code).collect::<Vec<Option<u8>>>());
         let ipkt = UInt64Array::from(self.flows.iter().map(|p| p.ipkt).collect::<Vec<Option<u64>>>());
         let ibyt = UInt64Array::from(self.flows.iter().map(|p| p.ibyt).collect::<Vec<Option<u64>>>());
         let smk = UInt8Array::from(self.flows.iter().map(|p| p.smk).collect::<Vec<Option<u8>>>());
@@ -223,8 +188,7 @@ impl FlowWriter {
         let outif = UInt16Array::from(self.flows.iter().map(|p| p.outif).collect::<Vec<Option<u16>>>());
         let sas = UInt32Array::from(self.flows.iter().map(|p| p.sas).collect::<Vec<Option<u32>>>());
         let das = UInt32Array::from(self.flows.iter().map(|p| p.das).collect::<Vec<Option<u32>>>());
-        let exid = UInt16Array::from(self.flows.iter().map(|p| p.exid).collect::<Vec<Option<u16>>>());
-        // let flowsrc = GenericStringArray::<i32>::from(vec![self.source_name.clone(); self.flows.len()]);
+        // let exid = UInt16Array::from(self.flows.iter().map(|p| p.exid).collect::<Vec<Option<u16>>>());
         let flowsrc = GenericStringArray::<i32>::from(self.flows.iter().map(|p| p.flowsrc.clone()).collect::<Vec<Option<String>>>());
 
         let batch = RecordBatch::try_new(
@@ -238,8 +202,8 @@ impl FlowWriter {
                 Arc::new(dp),
                 Arc::new(pr),
                 Arc::new(flg),
-                // Arc::new(icmp_type),
-                // Arc::new(icmp_code),
+                Arc::new(icmp_type),
+                Arc::new(icmp_code),
                 Arc::new(ipkt),
                 Arc::new(ibyt),
                 Arc::new(smk),
@@ -249,7 +213,7 @@ impl FlowWriter {
                 Arc::new(outif),
                 Arc::new(sas),
                 Arc::new(das),
-                Arc::new(exid),
+                // Arc::new(exid),
                 Arc::new(flowsrc),
                 ]
         ).unwrap();
@@ -258,7 +222,7 @@ impl FlowWriter {
 
     }
 
-    fn ch_query(&mut self, query: String) {
+    fn ch_query(&mut self, query: String) -> bool {
 
         debug!("CH query: {}", query.clone());
 
@@ -280,17 +244,27 @@ impl FlowWriter {
         cmd.arg("--query");
         cmd.arg(query);
     
-
-        // let output = match cmd.output() {
-        let _output = match cmd.spawn().expect("Error insert").wait_with_output() {
-            Err(e) => debug!("{:?}", e),
-            Ok(_output) => {
-                debug!("OK");
+        cmd.stderr(std::process::Stdio::piped());
+        cmd.stdout(std::process::Stdio::null());
+        let mut success = true;
+        match cmd.spawn().expect("Error query").wait_with_output() {
+            Err(e) => {
+                success = false;
+                debug!("{:?}", e)
+            }
+            Ok(output) => {
+                // debug!("OK -> {:?}", output.status.code().unwrap());
+                if output.status.code() != Some(0) {
+                    success = false;
+                    debug!("Some error when querying (exit code: {}, stderr={:?}", 
+                        output.status.code().unwrap(), 
+                        String::from_utf8(output.stderr));
+                }
             }
         };
-        // debug!("{:#?}", output);
+        success
     }
-    
+
     fn create_db_and_table(&mut self) {
 
         let ttlstr = match self.ch_ttl {
@@ -310,6 +284,8 @@ impl FlowWriter {
                 `dp` UInt16 DEFAULT 0,
                 `pr` Nullable(String),
                 `flg` LowCardinality(String),
+                `icmp_type` UInt8 DEFAULT 0,
+                `icmp_code` UInt8 DEFAULT 0,
                 `ipkt` UInt64,
                 `ibyt` UInt64,
                 `smk` UInt8,
@@ -319,7 +295,6 @@ impl FlowWriter {
                 `out` UInt16 DEFAULT 0,
                 `sas` UInt32 DEFAULT 0,
                 `das` UInt32 DEFAULT 0,
-                `exid` UInt16 DEFAULT 0,
                 `flowsrc` LowCardinality(String)
             )
             ENGINE = MergeTree
